# docs/reference/openapi.yaml
openapi: 3.1.0
info:
    title: durable-actors control plane API
    version: "1"
    description: |
        Manage durable-actors deployments, find actor hosts, and inspect activity.
        Authenticate backend requests with a bearer API key; never expose it to browsers.
        Project-scoped actor routes work in development and production. Run `durable-actors dev`
        and copy the printed connection settings into your backend's environment. The default
        local project ID is `local`; use it as `project_id` in requests.
servers:
    - url: http://127.0.0.1:7100
      description: Local server.
security:
    - ApiKey: []
...
paths:
...
    /v1/projects/{project_id}/deployment:
        parameters:
            - { $ref: "#/components/parameters/ProjectId" }
        put:
            operationId: registerDeployment
            tags: [Deployments]
            summary: Deploy actors
            description: Builds the supplied source image, replaces the current deployment, and restarts its actors while preserving saved state. Local development registers source from the filesystem with imageRef set to local.
            requestBody:
                required: true
                content:
                    application/json:
                        schema: { $ref: "#/components/schemas/Deployment" }
                        examples:
                            chat:
                                value:
                                    imageRef: im-customer-build
                                    workingDirectory: /customer
                                    actorEntrypoint: src/actors.ts
                                    secretRefs: []
            responses:
                "200": { $ref: "#/components/responses/Changed" }
                "400": { $ref: "#/components/responses/InvalidRequest" }
                "401": { $ref: "#/components/responses/Unauthenticated" }
                "413": { $ref: "#/components/responses/PayloadTooLarge" }
        get:
            operationId: getDeployment
            tags: [Deployments]
            summary: Get current deployment
            responses:
                "200":
                    description: Active deployment, without the contract.
                    content:
                        application/json:
                            schema: { $ref: "#/components/schemas/CurrentDeployment" }
                "400": { $ref: "#/components/responses/InvalidRequest" }
                "401": { $ref: "#/components/responses/Unauthenticated" }
                "404": { $ref: "#/components/responses/NotFound" }
                "500": { $ref: "#/components/responses/Internal" }
        delete:
            operationId: deleteDeployment
            tags: [Deployments]
            summary: Delete deployment
            description: Stops deployed actors and keeps their saved state.
            responses:
                "200": { $ref: "#/components/responses/Changed" }
                "400": { $ref: "#/components/responses/InvalidRequest" }
                "401": { $ref: "#/components/responses/Unauthenticated" }
                "500": { $ref: "#/components/responses/Internal" }
...
    /v1/projects/{project_id}/actors/{actor_name}/{actor_id}/find-actor:
        parameters:
            - { $ref: "#/components/parameters/ProjectId" }
            - { $ref: "#/components/parameters/ActorName" }
            - { $ref: "#/components/parameters/ActorId" }
        post:
            operationId: findActor
            tags: [Actors]
            summary: Find an actor host
            description: Resolves placement, starts a host if needed, and issues credentials for direct gRPC calls. Requires an active deployment.
            requestBody: { $ref: "#/components/requestBodies/FindActor" }
            responses:
                "200": { $ref: "#/components/responses/ActorTarget" }
                "400": { $ref: "#/components/responses/InvalidRequest" }
                "401": { $ref: "#/components/responses/Unauthenticated" }
                "409": { $ref: "#/components/responses/Conflict" }
                "413": { $ref: "#/components/responses/PayloadTooLarge" }
                "500": { $ref: "#/components/responses/Internal" }
                "503": { $ref: "#/components/responses/Unavailable" }
    /v1/projects/{project_id}/actors/{actor_name}/{actor_id}/find-websocket:
        parameters:
            - { $ref: "#/components/parameters/ProjectId" }
            - { $ref: "#/components/parameters/ActorName" }
            - { $ref: "#/components/parameters/ActorId" }
        post:
            operationId: findWebSocket
            tags: [Actors]
            summary: Find an authorized WebSocket URL
            description: Issues a fresh connection URL. Authorize the application user before requesting it. Requires an active deployment.
            requestBody: { $ref: "#/components/requestBodies/FindWebSocket" }
            responses:
                "200": { $ref: "#/components/responses/WebSocketGrant" }
                "400": { $ref: "#/components/responses/InvalidRequest" }
                "401": { $ref: "#/components/responses/Unauthenticated" }
                "409": { $ref: "#/components/responses/Conflict" }
                "413": { $ref: "#/components/responses/PayloadTooLarge" }
                "500": { $ref: "#/components/responses/Internal" }
                "503": { $ref: "#/components/responses/Unavailable" }
...
components:
    securitySchemes:
        ApiKey:
            type: http
            scheme: bearer
            description: Backend API key; never send it to browsers.
        SocketTicket:
            type: apiKey
            in: query
            name: key
            description: Included in websocketUrl. Keep the URL private.
    headers:
        NoStore:
            schema: { type: string, const: no-store }
    parameters:
...
        ProjectId:
            name: project_id
            in: path
            required: true
            schema: { type: string, pattern: "^[A-Za-z0-9._-]{1,64}$" }
        ActorName:
            name: actor_name
            in: path
            required: true
            schema: { type: string, pattern: "^[A-Za-z0-9._-]{1,48}$" }
        ActorId:
            name: actor_id
            in: path
            required: true
            schema: { type: string, pattern: "^[A-Za-z0-9._-]{1,128}$" }
    requestBodies:
        FindActor:
            required: true
            content:
                application/json:
                    schema: { $ref: "#/components/schemas/FindActorRequest" }
                    examples:
                        default: { value: {} }
                        regional: { value: { homeRegion: north-america-west } }
        FindWebSocket:
            required: true
            content:
                application/json:
                    schema: { $ref: "#/components/schemas/FindWebSocketRequest" }
                    examples:
                        socket: { value: { metadata: { userId: alice }, authorizationLifetimeMs: 900000 } }
    responses:
        Changed:
            description: Whether the deployment changed.
            content:
                application/json:
                    schema:
                        type: object
                        required: [changed]
                        properties:
                            changed: { type: boolean }
        ActorTarget:
            description: Owning host and RPC credentials. Timestamps use Unix milliseconds.
            headers:
                Cache-Control: { $ref: "#/components/headers/NoStore" }
            content:
                application/json:
                    schema: { $ref: "#/components/schemas/ActorTarget" }
        WebSocketGrant:
            description: Authorized WebSocket URL and deadlines. Timestamps use Unix milliseconds.
            headers:
                Cache-Control: { $ref: "#/components/headers/NoStore" }
            content:
                application/json:
                    schema: { $ref: "#/components/schemas/WebSocketGrant" }
...
        InvalidRequest:
            description: Invalid request (invalid_request).
            content:
                application/json:
                    schema: { $ref: "#/components/schemas/Error" }
        Unauthenticated:
            description: Invalid or missing API key (unauthenticated).
            content:
                application/json:
                    schema: { $ref: "#/components/schemas/Error" }
        NotFound:
            description: Not found (not_found).
            content:
                application/json:
                    schema: { $ref: "#/components/schemas/Error" }
        Conflict:
            description: Deployment or region conflict (conflict).
            content:
                application/json:
                    schema: { $ref: "#/components/schemas/Error" }
        PayloadTooLarge:
            description: Request exceeds 16 MiB (payload_too_large).
            content:
                application/json:
                    schema: { $ref: "#/components/schemas/Error" }
        Internal:
            description: Server failure (internal).
            content:
                application/json:
                    schema: { $ref: "#/components/schemas/Error" }
        Unavailable:
            description: Temporarily unavailable (unavailable).
            content:
                application/json:
                    schema: { $ref: "#/components/schemas/Error" }
...
    schemas:
...
        Deployment:
            type: object
            additionalProperties: false
            required: [imageRef, workingDirectory]
            properties:
                imageRef:
                    type: string
                    minLength: 1
                    maxLength: 255
                    description: Published Modal source image ID (im-...) for hosted deployments; local for development. Maximum 255 UTF-8 bytes.
                workingDirectory:
                    type: string
                    pattern: "^/"
                    description: Absolute project path inside the image. Maximum 1024 UTF-8 bytes.
                actorEntrypoint:
                    type: [string, "null"]
                    minLength: 1
                    default: null
                    description: Source path; null uses src/actors.ts. Maximum 1024 UTF-8 bytes.
                secretRefs:
                    type: array
                    maxItems: 16
                    default: []
                    items: { type: string, pattern: "^[A-Za-z0-9._-]{1,255}$" }
                    description: Secret names for deployed actors.
                contract:
                    oneOf:
                        - { $ref: "#/components/schemas/PublicActorContract" }
                        - { type: "null" }
                    default: null
                    description: Optional; generated for hosted deployments. Maximum 4 MiB.
        CurrentDeployment:
            type: object
            additionalProperties: false
            required: [imageRef, workingDirectory, actorEntrypoint, secretRefs]
            properties:
                imageRef: { $ref: "#/components/schemas/Deployment/properties/imageRef" }
                workingDirectory: { $ref: "#/components/schemas/Deployment/properties/workingDirectory" }
                actorEntrypoint: { $ref: "#/components/schemas/Deployment/properties/actorEntrypoint" }
                secretRefs: { $ref: "#/components/schemas/Deployment/properties/secretRefs" }
        PublicActorContract:
            type: object
            additionalProperties: false
            required: [version, actors]
            properties:
                version: { type: integer, const: 1 }
                actors:
                    type: array
                    items:
                        type: object
                        additionalProperties: false
                        required: [actorName, socket, rpc]
                        properties:
                            actorName: { type: string }
                            socket:
                                type: object
                                additionalProperties: false
                                required: [version, actorName, schema, emittable]
                                properties:
                                    version: { type: integer, const: 1 }
                                    actorName: { type: string }
                                    schema: { type: object, description: JSON Schema for socket types. }
                                    emittable: { type: array, uniqueItems: true, items: { type: string } }
                            rpc:
                                type: object
                                additionalProperties: false
                                required: [schema, methods]
                                properties:
                                    schema: { type: object, description: JSON Schema for RPC types. }
                                    methods: { type: array, items: { $ref: "#/components/schemas/RpcMethod" } }
        RpcMethod:
            type: object
            additionalProperties: false
            required: [name, parameters, result]
            properties:
                name: { type: string }
                parameters:
                    type: array
                    items:
                        type: object
                        additionalProperties: false
                        required: [name, type, optional, rest]
                        properties:
                            name: { type: string }
                            type: { $ref: "#/components/schemas/RpcType" }
                            optional: { type: boolean }
                            rest: { type: boolean }
                result:
                    oneOf:
                        - type: object
                          additionalProperties: false
                          required: [kind]
                          properties:
                              kind: { type: string, const: void }
                        - type: object
                          additionalProperties: false
                          required: [kind, type]
                          properties:
                              kind: { type: string, const: value }
                              type: { $ref: "#/components/schemas/RpcType" }
        RpcType:
            type: object
            additionalProperties: false
            required: ["$ref"]
            properties:
                $ref: { type: string, description: Reference into rpc.schema. }
        FindActorRequest:
            type: object
            additionalProperties: false
            properties:
                homeRegion: { $ref: "#/components/schemas/HomeRegion" }
        FindWebSocketRequest:
            type: object
            additionalProperties: false
            required: [metadata]
            properties:
                homeRegion: { $ref: "#/components/schemas/HomeRegion" }
                metadata: { description: Required connection metadata; may be null. Maximum 64 KiB of JSON. }
                authorizationLifetimeMs: { type: integer, minimum: 1000, maximum: 86400000, default: 900000 }
        HomeRegion:
            type: [string, "null"]
            pattern: "^[a-z0-9._-]{1,64}$"
            description: Region for new actors; cannot change an existing actor's region. Omit or use null for automatic placement.
        ActorTarget:
            type: object
            required: [homeRegion, route, token, ownerEpoch, expiresAtMs]
            properties:
                homeRegion: { type: string }
                route: { type: string, format: uri, description: HTTP(S) origin of the actor gRPC host. }
                token: { type: string, description: Bearer token for actor RPCs. }
                ownerEpoch: { type: integer, minimum: 0 }
                expiresAtMs: { type: integer, format: int64 }
        WebSocketGrant:
            type: object
            required: [homeRegion, websocketUrl, connectByMs, authorizedUntilMs]
            properties:
                homeRegion: { type: string }
                websocketUrl: { type: string, format: uri }
                connectByMs: { type: integer, format: int64, description: Connect before this time. }
                authorizedUntilMs: { type: integer, format: int64, description: Connection expires at this time. }
...
        Error:
            type: object
            required: [error]
            properties:
                error:
                    type: object
                    required: [code, message]
                    properties:
                        code: { type: string }
                        message: { type: string }